requirements.txt
flask==2.1.2
requests==2.27.1
pytest==7.1.2
coverage==6.4.1
gunicorn==20.1.0
python-dotenv==0.20.2